+++ Runtime/World/Components/RigidBody.cpp
@@ -122,7 +122,7 @@ namespace Spartan
         Activate();
     }
 
-    void RigidBody::OnTick(float delta_time)
+    void RigidBody::OnTick(double delta_time)
     {
         // When the rigid body is inactive or we are in editor mode, allow the user to move/rotate it
         if (!IsActivated() || !m_context->m_engine->EngineMode_IsSet(Engine_Game))
